backend/pyproject.toml
[tool.poetry]
name = "django-backend"
version = "0.1.0"
description = ""
authors = ["Piotr Truty"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
mysqlclient = "^2.2.5"
django-rest-framework = "^0.1.0"
boto3 = "^1.35.50"
requests = "^2.32.3"
slack-bolt = "^1.21.2"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
